removed plotting test and merged width_systematics with systematics.py
